Self-cleaning obsolete network interfaces in FreeBSD plugin
